Bug fixes to utils + restructured utils
